fix: keep the same Redis client open to reduce the number of open connections
