+++ src/server.ts
@@ -5,7 +5,6 @@ const server = app.listen(PORT, () => {
   console.green(`🟢 Server started on port ${PORT}`);
 
   /* Used to test server build */
-  // @ts-expect-error Nodejs process override
   if (process.env.NODE_ENV === 'build') {
     console.blue('🧪 Testing server build');
     setTimeout(() => {
